Fixed reputation issues in migrations.

+++ db/migrate/20120729163249_create_chat_messages.rb
@@ -14,7 +14,9 @@ class CreateChatMessages < ActiveRecord::Migration
       t.timestamps
     end
 
-    remove_column :profiles,:reputation
+    if column_exists? :profiles,:reputation
+      remove_column :profiles,:reputation
+    end
     add_column :profiles,:reputation,:integer,default: 0
   end
 end
